ARROW-578: [C++] Add -DARROW_CXXFLAGS=... option to make CMake more consistent
I've had issues on CMake 2.8.x with `-DCMAKE_CXX_FLAGS=$MY_CXXFLAGS` not passing on the flags to the compiler. But it seems to work properly in our Travis CI setup, so go figure.

Some Google searches seem to confirm this is a known issue, and having a specific "user flags" option is a way around it. We just did the same thing in parquet-cpp.

Author: Wes McKinney <wes.mckinney@twosigma.com>

Closes #348 from wesm/ARROW-578 and squashes the following commits:

1103bed [Wes McKinney] Use ARROW_CXXFLAGS in Travis CI
086d643 [Wes McKinney] Add -DARROW_CXXFLAGS=... option to make CMake behavior more consistent across versions