ARROW-5544: [Archery] Don't return non-zero on regressions
The return code should be used to return when the command did not
execute successfully. In this case, the command did run without issues,
we leave to the calling user to decide what to do with the data.

This will simplify integration with ursabot CI.

Author: François Saint-Jacques <fsaintjacques@gmail.com>

Closes #4512 from fsaintjacques/ARROW-5544-archery-benchmark-diff-return-code and squashes the following commits:

b22231356 <François Saint-Jacques> ARROW-5544:  don't return non-zero on regressions