ARROW-7460: [Rust] Improve some kernel performance
This removes the null checks on arithmetic and comparison kernels, in order to simplify the kernels and allow for LLVM to autovectorise the remaining simple loops.

Although `for loop`s are often slower than iterators (due to bounds checks), we do not perform bounds checking when accessing array values.
This results in performance being as good as what we might get with iterator support.

Another observation from my benchmarking is that the kernels are now as fast as their SIMD equivalents (on my 2012-13 Haswell).

Closes #6086 from nevi-me/ARROW-7460 and squashes the following commits:

4e54e72f3 <Neville Dipale> fix feature-gated imports
a434b725b <Neville Dipale> update flight dependencies to fix build failures
04ada4380 <Neville Dipale> ARROW-7460:  Improve some kernel performance

Authored-by: Neville Dipale <nevilledips@gmail.com>
Signed-off-by: Neville Dipale <nevilledips@gmail.com>