ARROW-7723: [Python] Triage untested functional regression when converting tz-aware timestamp inside struct to pandas/NumPy format
This is definitely a hack so I left notes everywhere and explained the problem in detail on the ARROW-7723 JIRA. This reverts behavior in this particular case to 0.15.1 but there is an inconsistency in the data types returned by `struct<f0: timestamp[us]>` versus `struct<f0: timestamp[ns]>`. The former returns a dict with `datetime.datetime` values while the latter has `int64` values (because nanoseconds can't be safely coerced to `datetime.datetime` in general).

It seems undesirable at the moment to have any downstream user depending on one behavior or the other at the moment but leaving the behavior unchanged for now (given that many systems -- e.g. Spark SQL -- don't have nanoseconds) seems acceptable.

Closes #6322 from wesm/ARROW-7723 and squashes the following commits:

05ab21bb9 <Wes McKinney> pin pandas 0.25.3
217897717 <Wes McKinney> Add longer code comment explaining the hack
7142699da <Wes McKinney> Gross hack to avoid unwanted nanosecond promotion with tz-aware timestamp types
123d4b81e <Wes McKinney> Add failing unit test

Authored-by: Wes McKinney <wesm+git@apache.org>
Signed-off-by: Krisztián Szűcs <szucs.krisztian@gmail.com>