ARROW-8536: [Rust] [Flight] Check in proto file, conditional build if file exists
When a user compiles the `flight` crate, a `build.rs` script is invoked. This script recursively looks for the `format/Flight.proto` path. A user might not have that path, as they would not have cloned the arrow repository, and as such, the build fails.

This change:

a) checks if the `../../format/Flight.proto` path exists, and only builds if the file exists, and has been changed.
b) checks in the proto file into the `src/` directory, changing the default location from an opaque directory in the `target/{debug|release}/build-xxx` folder.

The rationale for checking in the proto file is that if the file is not rebuilt, `flight` users are still able to access the file. It's also beneficial for users to easily view the file, and better understand how the generated code looks like.

Closes #7018 from nevi-me/ARROW-8536

Authored-by: Neville Dipale <nevilledips@gmail.com>
Signed-off-by: Neville Dipale <nevilledips@gmail.com>