ARROW-9261: [Python] Fix CA certificate lookup with S3 filesystem on manylinux
The AWS SDK on manylinux packages uses a custom-compiled libcurl that is configured for the certificates of the build system.  However, there's no standard location for CA certificates on Linux, so we must instead configure the filesystem layer at runtime with Python's own TLS configuration values (which are assumed to be correct, otherwise Python itself couldn't connect to TLS hosts).

Closes #7580 from pitrou/ARROW-9261-manylinux-s3-certificates

Authored-by: Antoine Pitrou <antoine@python.org>
Signed-off-by: Antoine Pitrou <antoine@python.org>