ARROW-8735: [Rust] [Parquet] Allow arm 32 to use soft hash implementation
Minimal change to allow this to be used on arm 32 targets.

A better solution may be to utilise the crc32 instruction ala the optimised intel implementation but this is better than nothing. There should probably be a crate we could just import the hash implementation from rather than having to repeat optimisations everywhere.

Closes #8338 from andrew-stevenson-frequenz/issue-8735

Authored-by: Andrew Stevenson <andrew.stevenson@frequenz.com>
Signed-off-by: Andy Grove <andygrove@nvidia.com>