ARROW-10289: [Rust] Read dictionaries in IPC streams
We were reading dictionaries in the file reader, but not in the stream reader.
This was a trivial change, as we needed to add the dictionary to the stream when we encounter it, and then read the next message until we reach a record batch.

I tested with the 0.14.1 golden file, I'm going to test with later versions (1.0.0-littleendian) when I get to `arrow::ipc::MetadataVersion::V5` support, hopefully soon.

Closes #8450 from nevi-me/ARROW-10289

Authored-by: Neville Dipale <nevilledips@gmail.com>
Signed-off-by: Jorge C. Leitao <jorgecarleitao@gmail.com>