ARROW-9361 [Rust] Move array types into their own modules
This PR splits the `array.rs` into multiple modules, so that it is a bit easier to navigate in the project, particularly `array.rs`. Semantically, there should be no difference, and this change is also backward compatible, public items are only exposed in `mod.rs`.

This change was initially proposed by @paddyhoran on ARROW-9361, and this is just a proposed implementation.

Some notes:

* Each commit is a different array
* I named the files `array_[type]` to distinguish from other modules that do not contain arrays.
* `null` and `Union` were not renamed
* Some functions were moved around / incorporated in `impl` since they were used in multiple places.

Closes #8543 from jorgecarleitao/clean_slice

Authored-by: Jorge C. Leitao <jorgecarleitao@gmail.com>
Signed-off-by: Neville Dipale <nevilledips@gmail.com>