ARROW-10552: [Rust] Removed un-used Result
This PR is a minor simplification of the code base around `Buffer::reserve`. The idea is that this function is infalible (it just returns `Ok(...)`) and therefore does not require to return a `Result`. Since this function is on a very low level, it leads to the removal of the `Result` in some places.

Closes #8636 from jorgecarleitao/clean_result

Authored-by: Jorge C. Leitao <jorgecarleitao@gmail.com>
Signed-off-by: alamb <andrew@nerdnetworks.org>