ARROW-10542: [C#][Flight] Add beginning on flight code for net core
I closed the previous PR and opened this one, since the original one did not use InternalsVisibleTo and required a version bump on the Apache.Arrow project from netstandard1.3 to netstandard1.5.

This adds basic support for both a flight server and a flight client for Net Core.
Sorry for the massive PR, but I wanted to have some basic functionality in place with atleast some tests that use the put and get flow before creating a PR.
I hope that this PR can help create some discussion on how the interfaces should look etc, and if this looks like an accetable interface for flight in net core. It tries to mimic as much as possible the original gRPC net core interface, but mapping the classes from the network protocol to the C# classes.

This implementation uses InternalsVisibleTo, to hinder a bump of .netstandard version from 1.3 to 1.5.
So the Apache.Arrow project still uses netstandard1.3.
All flight code is in a seperate project Apache.Arrow.Flight

This also required changing build version from 2.2 to 3.0

The code does not include:

* Handshake - the reason is that AspNetCore contains features already for authentication/authorization for gRPC. Can be added later ofcourse.
* DoExchange - I feel that more feedback/discussion is required before DoExchange can be implemented.

Note:
**This has been fixed**
Sourcelink did not work when using grpc.tools to have code compilation in the build step. So I had to generate the grpc code manually for sourcelink to work. This means that there are alot of extra code in this PR that are auto generated.

cc: @eerhardt
Looks like you are the most active for C#, would be great to get some feedback if this is similar to how you would implement it, or if major changes are required I am ofcourse up for that as well.

Closes #8694 from Ulimo/master

Lead-authored-by: Östman Alexander <alexander.ostman@sweco.se>
Co-authored-by: Ulimo <alexander.ostman@hotmail.com>
Signed-off-by: Eric Erhardt <eric.erhardt@microsoft.com>