ARROW-10674: [Rust] Fix BigDecimal to be little endian; Add IPC Reader/Writer for Decimal type to allow integration tests
This is a follow up to #8640

Currently, there is a first working IPC reader/writer test using data from `testing/arrow-ipc-stream/integration/0.14.1/generated_decimal.arrow_file`

However, this lead me to discover that my first decimal type implementation is wrong, in that it uses BigEndian, whereas this is parquet specific and therefore should not be used in arrow/array and so on. I will try to address this in this PR as well.

Closes #8784 from sweb/rust-decimal-ipc

Authored-by: Florian Müller <florian@tomueller.de>
Signed-off-by: Jorge C. Leitao <jorgecarleitao@gmail.com>