ARROW-10943: [Rust][Parquet] Always init new RleDecoder
Part of the removal of specialisation makes the RleDecoder remain
instanticated. This is done to lose some allocations but right now
appears to leave a decoder in a semi configured state that can mis-read
RLE packed data.

As such we just init a new one each time the parent decoder has data set
on it in a fashion to the pre refactored code.

Closes #8993 from GregBowyer/arrow-10943

Authored-by: Greg Bowyer <gbowyer@fastmail.co.uk>
Signed-off-by: Jorge C. Leitao <jorgecarleitao@gmail.com>