ARROW-11355: [Rust] Aligned Date DataType with specification.
The arrow spec states that `Date32` represents a date in number of days, and `Date64` represents a date in milliseconds.

However, our current implementation allows `Date32(Milliseconds)` and `Date64(Days)`.

This PR fixes this by removing the `DateUnit` altogether, i.e. `Date32(_) -> Date32` and `Date64(_) -> Date64`. This choice is based on how this is offered on pyarrow's API: https://arrow.apache.org/docs/python/generated/pyarrow.date32.html

Please note that this has a major backward incompatibility to all code that uses this datatype. Unfortunately, there is no way to solve this out of spec without an incompatible change. :(

Closes #9298 from jorgecarleitao/simplify_date

Authored-by: Jorge C. Leitao <jorgecarleitao@gmail.com>
Signed-off-by: Jorge C. Leitao <jorgecarleitao@gmail.com>