ARROW-11442: [Rust] Expose datetime conversion logic independently
This is a small refactor that exposes the functions used to convert dates and times into `NaiveDate`, `NaiveTime` and `NaiveDateTime`. They are currently hidden beind the `PrimitiveArray<ArrowTemporalType>::value_as_time`, which makes them difficult to use when the transformation is to be applied in SIMD, or when we want to reproduce the operation for a single scalar value (not in an arrow array).

Closes #9378 from jorgecarleitao/conversions

Authored-by: Jorge C. Leitao <jorgecarleitao@gmail.com>
Signed-off-by: Andrew Lamb <andrew@nerdnetworks.org>