ARROW-10457: [CI] Fix Spark integration tests with branch-3.0
This adds an additional docker-compose task that will test a Spark branch with the latest PyArrow, instead of the default which also builds Spark with the latest Arrow Java. This is so existing Spark releases with fixed Arrow Java dependencies can still be tested with current PyArrow changes to ensure there is still compatibility. The nightly test against Spark branch-3.0 is changed to use this new task.

Closes #9210 from BryanCutler/ci-spark-branch-3.0-ARROW-10457

Authored-by: Bryan Cutler <cutlerb@gmail.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>