ARROW-11627: [Rust] Make allocator be a generic over type T
The background and rational for this is described [here](https://github.com/jorgecarleitao/arrow2/tree/proposal); the idea is that this is groundwork to make our buffers typed, so that we can start introducing strong typing in the crate.

This change is backward incompatible:

1. Our allocator is now a generic over type `T: NativeType`, which implies that we can now allocate certain types.
2. The allocator moved from `memory` to a new module `alloc` (inspired after `std::alloc`).

Necessary steps to migrate existing code:

1. `use arrow::memory` -> `use arrow::alloc`
2. `memory::allocate_aligned(...)` -> `alloc::allocate_aligned::<u8>(...)`

Note how `NativeType` contains `to_le_bytes`; we will use this method for IPC, where we need to serialize buffers with a specific endianess. This is ground work to enable multiple endianesses support

Closes #9495 from jorgecarleitao/alloc_t

Authored-by: Jorge C. Leitao <jorgecarleitao@gmail.com>
Signed-off-by: Neville Dipale <nevilledips@gmail.com>